requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose
cloudinary
sentence-transformers
faiss-cpu
openai
python-multipart
celery
redis
pypdf